feat: add trigger metric

+++ internal/trigger/worker.go
@@ -27,6 +27,7 @@ import (
 	"github.com/linkall-labs/vanus/internal/trigger/errors"
 	"github.com/linkall-labs/vanus/internal/trigger/trigger"
 	"github.com/linkall-labs/vanus/observability/log"
+	"github.com/linkall-labs/vanus/observability/metrics"
 	ctrlpb "github.com/linkall-labs/vanus/proto/pkg/controller"
 	metapb "github.com/linkall-labs/vanus/proto/pkg/meta"
 )
@@ -165,14 +166,20 @@ func (w *worker) AddSubscription(ctx context.Context, subscription *primitive.Su
 		return err
 	}
 	w.addTrigger(subscription.ID, t)
+	metrics.TriggerGauge.WithLabelValues(subscription.EventBus, w.config.IP).Inc()
 	return nil
 }
 
 func (w *worker) RemoveSubscription(ctx context.Context, id vanus.ID) error {
 	w.lock.Lock()
 	defer w.lock.Unlock()
+	t, exist := w.getTrigger(id)
+	if !exist {
+		return nil
+	}
 	_ = w.stopSubscription(ctx, id)
 	w.deleteTrigger(id)
+	metrics.TriggerGauge.WithLabelValues(t.GetSubscription(ctx).EventBus, w.config.IP).Dec()
 	return nil
 }
